Walker_Lake

A dataset with geocoded observations of a series of variables originally used for teaching geostatistics in Isaaks and Srivastava's An Introduction to Geostatistics. The variables are as follows:

Usage

data(Walker_Lake)

Format

A dataset with 470 rows and 6 variables

Source

https://books.google.ca/books/about/Applied_Geostatistics.html?id=vC2dcXFLI3YC&redir_esc=y

Details

ID. object ID of variable (1–470)

X. false x-coordinates (8–251)

Y. false y-coordinates (8–291)

V. quantitative variable (0–1528.1)

U. quantitative variable (0–5190.1)

T. factor variable (1–2)