isdas: Companion Package for Book An Introduction to Spatial Data Analysis and Statistics

This package contains data sets, custom functions, and document templates to work with the book An Introduction to Spatial Data Analysis and Statistics: A Course in R.

See also

Author

Maintainer: Antonio Páez paezha@mcmaster.ca (ORCID)

Authors:

Rajveer Ubhi ubhir20@gmail.com

Anastassios Dardas