A funcion for creating an interactive map of local Moran's I statistics

This function plots a map with the results of local Moran's I statistics, based on spatial polygons, a listw object, and a variable with an ID.

Usage

localmoran.map(p = p, listw = listw, VAR = VAR, by = by)

Arguments

p

A simple features sf object

listw

A listw object (see spdep)

VAR

The name of a variable

by

The name of a key to join the local Moran's I statistics to the sf object

Examples

# Create a map of local Moran's I statistics for population density
#
data(Hamilton_CT)
# Obtain a listw object for the contiguities. First obtain the neighbors:
Hamilton_CT.nb <- spdep::poly2nb(Hamilton_CT)
# Based on the neighbors, obtain a listw object:
Hamilton_CT.w <- spdep::nb2listw(Hamilton_CT.nb)

localmoran.map(Hamilton_CT, Hamilton_CT.w, "POP_DENSITY", "TRACT")
#> Error in stopifnot(!inherits(x, "sf"), !missing(sf_column_name), !missing(agr)): ℹ In argument: `SMA = lag.listw(listw, Z)`.
#> Caused by error in `lag.listw()`:
#> ! could not find function "lag.listw"