A funcion for creating an interactive map of local Moran's I statistics

This function plots a map with the results of local Moran's I statistics, based on spatial polygons, a listw object, and a variable with an ID.

Usage

localmoran.map(p = p, listw = listw, VAR = VAR, by = by)

Arguments

p

A simple features sf object

listw

A listw object (see spdep)

VAR

The name of a variable

by

The name of a key to join the local Moran's I statistics to the sf object

Examples

# Create a map of local Moran's I statistics for population density
#
# Obtain a listw object for the contiguities. First obtain the neighbors:
Hamilton_CT.nb <- spdep::poly2nb(as(Hamilton_CT, "Spatial"))
#> Error in .class1(object): object 'Hamilton_CT' not found
# Based on the neighbors, obtain a listw object:
Hamilton_CT.w <- spdep::nb2listw(Hamilton_CT.nb)
#> Error in spdep::nb2listw(Hamilton_CT.nb): object 'Hamilton_CT.nb' not found

localmoran.map(Hamilton_CT, Hamilton_CT.w, "POP_DENSITY", "TRACT")
#> Error in transmute(p, key = p[[by]], Z = (p[[VAR]] - mean(p[[VAR]]))/var(p[[VAR]]),     SMA = lag.listw(listw, Z), Type = factor(ifelse(Z < 0 & SMA <         0, "LL", ifelse(Z > 0 & SMA > 0, "HH", "HL/LH")))): object 'Hamilton_CT' not found